requirements.txt
gymnasium==1.2.3
hydra-core==1.3.2
matplotlib==3.10.8
mysql_connector_repackaged==0.3.1
numpy==2.4.0
omegaconf==2.3.0
pandas==2.3.3
seaborn==0.13.2
torch==2.9.1+cu126
